back/requirements.txt
asgiref==3.2.7
Django==3.0.4
djangorestframework==3.11.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
django-cors-headers==3.2.1
requests==2.23.0